requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.2.0
python-dotenv==1.0.1
django-filter==23.5
django-storages==1.14.2
boto3==1.34.34
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0 